feat: PR3-3 C-3 searchDoubleMoveSuperAction 2 手目タダ捨て除外 + hasHangingPiece 共通化
## 背景
PR2 (#224) 残課題: double_move カード使用後の 2 手指しで「2 手目に相手駒の前へ
自駒を打つ → 次相手手番でタダ取り」する組合せが評価で除外されていなかった。
通常 move のタダ捨ては blunder guard で防止されているが、super-action 内部探索
(searchDoubleMoveSuperAction) には未適用だった。本 C-3 で解消。

## 変更内容

### src/lib/shogi/ai/blunder-guard.ts
- hasHangingPiece を engine.ts から移動 + export
  (search.ts double_move でも利用するため共通化、循環依存回避)
- BLUNDER_PIECE_VALUES 定数も同時に移動 (hasHangingPiece の依存)
- isSquareAttackedByFast / getLeastAttackerValue を import (依存解決)

### src/lib/shogi/ai/engine.ts
- hasHangingPiece / BLUNDER_PIECE_VALUES を削除 (blunder-guard.ts へ移動)
- blunder-guard.ts から hasHangingPiece を import
- 未使用になった isSquareAttackedByFast / getLeastAttackerValue import を除去

### src/lib/shogi/ai/search.ts
- searchDoubleMoveSuperAction に excludeTadasute 引数追加 (デフォルト false で
  既存呼び出し互換)
- 2 手目評価ループ:
  - bestScoreIgnoringTadasute をフォールバック用に維持
  - excludeTadasute=true 時に hasHangingPiece チェックで「タダ捨て組合せ」を skip
  - 全組合せが除外で消えた場合のフォールバック (= bestScore が NEG_INF のまま) は
    bestScoreIgnoringTadasute を返す (R-3 計画 md 参照、空集合で double_move が
    過剰減点されないよう保護)
- 呼び出し側 (evaluateAction の playCard double_move 分岐 / C-1 の
  evaluateActionWithLookahead) に excludeTadasute 伝播

## 振る舞いキープ
- excludeTadasute=false (デフォルト) なら既存挙動完全互換
- engine.ts root の呼び出しは applyTadasuteGuard を渡しており、本 C-3 で
  double_move 2 手目にも自然に効くようになる

## 検証
- npm run typecheck: クリーン
- npm run lint: 0 errors (21 既存 warning)
- npm run test:ci: 491 passed / 9 skipped (退化なし、double_move 既存テストも green)

対象 Issue #193 PR3-3 計画 md C-3 (PR2 残課題解消)
